requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.1
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
plotly>=5.18.0
requests>=2.31.0
nltk>=3.8.1
lxml>=5.1.0
html5lib>=1.1

# ===== Hybrid Database Architecture Dependencies =====
# Transactional Database (MySQL)
SQLAlchemy>=2.0.0
pymysql>=1.1.0

# AWS DynamoDB (NoSQL high-concurrency cache)
boto3>=1.26.0
botocore>=1.29.0

# Redis (distributed cache)
redis>=5.0.0

# Environment variable management
python-dotenv>=1.0.0

# Analytics Database (ClickHouse)
clickhouse-driver>=0.2.0  # ClickHouse real-time analytics

# Data processing and transformation
pyarrow>=12.0.0  # For efficient data transfer

# Logging and monitoring
python-json-logger>=2.0.7

# Test dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional AI client for sentiment analysis
openai>=1.0.0

